Refactor download code to separate function and allow shorter call for one sequence
This PR refactors the downloading code into a separate function in
order to resolve compiler warnings related to not handling the
return value of system (in a rather heavy-handed way). It also
includes a new two-argument calling interface, which duplicates
the same index for both the minimum and maximum as a shorthand.
I also added code to swap the indices if they're in the wrong
order.