Refactor file downloading and opening
This refactors file downloads into a new function, also resolving
the compiler warnings about ignoring the return value of `system`.
It no longer tries to work with partial ELF files, failing immediately
if curl returns with a partial download failure.

This was based on #11, as it was easier to apply the indentation
changes on the updated code.